requirements.txt
tensorflow-cpu==2.8.0
pillow==9.0.1
numpy==1.22.3
gunicorn==20.0.4
urllib3==1.26.9
flask==2.0.3
werkzeug==2.0.3
